+++ frontend-vue/src/components/common/AppHeader.vue
@@ -1,7 +1,7 @@
 <template>
     <v-app-bar color="light-blue">
         <v-app-bar-title class="font-weight-bold">
-            Employee ERP System
+            Employee ERP App
         </v-app-bar-title>
     </v-app-bar>
 </template>
