+++ frontend-vue/src/config/index.js
@@ -1,5 +1,5 @@
 export default {
     env: import.meta.env.VITE_ENV,
     port: import.meta.env.VITE_PORT,
-    apiURL: import.meta.env.VITE_API_URL
+    apiUrl: import.meta.env.VITE_API_URL
 };
