+++ frontend-vue/src/services/axios.js
@@ -3,7 +3,7 @@ import axios from 'axios';
 import config from '@/config';
 
 const axiosInstance = axios.create({
-    baseURL: config.apiURL,
+    baseURL: config.apiUrl,
     withCredentials: true
 });
