+++ frontend-vue/src/stores/user.js
@@ -1,6 +1,6 @@
 import { defineStore } from 'pinia';
 
-import axios from '@/services/api';
+import axios from '@/services/axios';
 
 export const useUserStore = defineStore('user', {
     state: () => ({
